[FEI-3474].1 Use modules for webpack
This replaces use of the webpack function with use of the modules array,
allowing us to reduce the bytes we need in our built loadables.

This also deletes the babel and webpack plugins herein since we have
custom versions of those at this point and the ones here just would
not work for us. No point having old code laying around.